requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
reportlab==4.0.7
python-multipart==0.0.6
openai>=1.0.0
openpyxl>=3.1.0

# Development / testing
pytest>=7.4.0
httpx>=0.25.0
